+++ docs/changelog.md
@@ -10,6 +10,10 @@ Requires libvips v8.10.5
 
 * Remove experimental status from `heif` output, defaults are now AVIF-centric.
 
+* Allow negative top/left offsets for composite operation.
+  [#2391](https://github.com/lovell/sharp/pull/2391)
+  [@CurosMJ](https://github.com/CurosMJ)
+
 * Ensure all platforms use fontconfig for font rendering.
   [#2399](https://github.com/lovell/sharp/issues/2399)